ZOD everything and working end to end db

+++ lib/server/SQLite.ts
@@ -1,6 +1,7 @@
 
 import path from "path";
 import { DatabaseSync } from 'node:sqlite';
+import { Task } from "@/lib/server/db";
 
 //process.env.NODE_ENV === "development" ? "db.dev.json" : "db.json";
 
@@ -17,22 +18,22 @@ database.exec('PRAGMA foreign_keys = on')
 database.exec(`CREATE TABLE IF NOT EXISTS users(
     username text UNIQUE NOT NULL,
     id text PRIMARY KEY,
-    data BLOB,
-    nonce BLOB);
+    data text,
+    nonce text);
     
   CREATE TABLE IF NOT EXISTS user_credentials(
     userId text PRIMARY KEY,
     passwordHash text NOT NULL,
-    wrappedDataKey text NOT NULL,
-    salt text NOT NULL,
-    nonce text NOT NULL,
+    wrappedDataKey BLOB NOT NULL,
+    salt BLOB NOT NULL,
+    nonce BLOB NOT NULL,
     foreign key(userId) references users(id) ON DELETE cascade);
     
   Create TABLE IF NOT EXISTS tasks(
       id text PRIMARY KEY,
       userId text NOT NULL ,
-      data BLOB,
-      nonce BLOB, 
+      data text,
+      nonce text, 
       foreign key(userId) references users(id) ON DELETE cascade
   );
   
@@ -54,7 +55,8 @@ database.exec(`CREATE TABLE IF NOT EXISTS users(
 
 
 export const authSql = {
-  getUser: database.prepare(`SELECT * FROM users u join user_credentials c on u.id = c.userId WHERE u.username = ? `),
+  getUser: database.prepare(`SELECT u.username, u.id, c.passwordHash, c.salt, c.wrappedDataKey, c.nonce
+     FROM users u join user_credentials c on u.id = c.userId WHERE u.username = ? `),
   getUserSalt: database.prepare(`SELECT c.salt FROM users u join user_credentials c on u.id = c.userId WHERE u.username = ?  `),
   createUserId: database.prepare(`INSERT INTO users(username, id) values(?,?)`),
   createUserCred: database.prepare(`INSERT INTO user_credentials(userId, passwordHash, wrappedDataKey, salt, nonce) values(?,?,?,?,?)`),
@@ -65,7 +67,7 @@ export const authSql = {
   generateFamily: database.prepare(`INSERT INTO token_family(idExp,userId) values(?,?)`),
   generateToken: database.prepare(`INSERT INTO refresh_tokens(id, isRevoked, familyId) values(?,?,?)`),
 
-  getToken: database.prepare(`SELECT r.isRevoked, t.id as familyId, t.idExp, t.userId as familyExp 
+  getToken: database.prepare(`SELECT r.isRevoked, t.id as familyId, t.idExp as idExp , t.userId 
     from refresh_tokens r  JOIN token_family t ON r.familyId = t.id 
     where r.id = ?`),
 
@@ -77,11 +79,11 @@ export const authSql = {
 
 export const dataSql = {
 
-  getTasks: database.prepare(`SELECT * FROM tasks where userId = ?`),
-  getTask: database.prepare(`SELECT * FROM tasks where id= ? and userId = ?`),
-  createTask: database.prepare(`INSERT INTO tasks(id, userId,data, nonce) values(?,?,?,?)`),
-  updateTask: database.prepare(`UPDATE tasks SET data = ? and nonce = ? where id = ? and userId = ?`),
-  deleteTask: database.prepare(`DELETE FROM tasks where id = ? and userId = ?`),
+  getTasks: database.prepare(`SELECT id, data, nonce FROM tasks where userId = @userId`),
+  getTask: database.prepare(`SELECT * FROM tasks where id= @id and userId = @userId`),
+  createTask: database.prepare(`INSERT INTO tasks(id, userId,data, nonce) values(@id,@userId,@data,@nonce)`),
+  updateTask: database.prepare(`UPDATE tasks SET data = @data, nonce = @nonce where id = @id and userId = @userId`),
+  deleteTask: database.prepare(`DELETE FROM tasks where id = @id and userId = @userId`),
   getUserData: database.prepare(`SELECT data, nonce FROM users where id = ?`),
-  saveUserData: database.prepare(`UPDATE users SET data = ?, nonce = ? where id = ? `)
+  saveUserData: database.prepare(`UPDATE users SET data = @data, nonce = @nonce where id = @userId `)
 }
